requirements.txt
discord.py
TenGiphPy
